Move saml strategy out of routes
